Question 1/average-calculator/src/requirements.txt
Django~=5.1.2
djangorestframework~=3.14.0
requests~=2.28.1